+++ docs/conf.py
@@ -73,7 +73,14 @@ opamp_dirs = [
     os.path.abspath("/".join([opamp, f, "src"]))
     for f in listdir(opamp)
     if isdir(join(opamp, f))
+
+util = "../util"
+util_dirs = [
+    os.path.abspath("/".join([util, f, "src"]))
+    for f in listdir(util)
+    if isdir(join(util, f))
 ]
+
 sys.path[:0] = (
     exp_dirs
     + instr_dirs
@@ -82,6 +89,7 @@ sys.path[:0] = (
     + prop_dirs
     + resource_dirs
     + opamp_dirs
+    + util_dirs
 )
 
 # -- Project information -----------------------------------------------------
@@ -131,6 +139,7 @@ intersphinx_mapping = {
         None,
     ),
     "redis": ("https://redis.readthedocs.io/en/latest/", None),
+    "fsspec": ("https://filesystem-spec.readthedocs.io/en/latest/", None),
 }
 
 # http://www.sphinx-doc.org/en/master/config.html#confval-nitpicky