+++ instrumentation-genai/opentelemetry-instrumentation-vertexai/tests/test_chat_completions.py
@@ -1,22 +1,7 @@
 from __future__ import annotations
 
-import asyncio
-from typing import (
-    Any,
-    Callable,
-    Generator,
-    Protocol,
-    TypeVar,
-)
-
 import pytest
 from google.api_core.exceptions import BadRequest, NotFound
-from google.auth.aio.credentials import (
-    AnonymousCredentials as AsyncAnonymousCredentials,
-)
-from google.cloud.aiplatform.initializer import _set_async_rest_credentials
-from typing_extensions import Concatenate, ParamSpec
-from vcr import VCR
 from vertexai.generative_models import (
     Content,
     GenerationConfig,
@@ -38,39 +23,45 @@ from opentelemetry.sdk.trace.export.in_memory_span_exporter import (
 from opentelemetry.trace import StatusCode
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr()
 def test_generate_content(
     span_exporter: InMemorySpanExporter,
     log_exporter: InMemoryLogExporter,
-    generate_content: GenerateContentFixture,
+    generate_content: callable,
     instrument_with_content: VertexAIInstrumentor,
 ):
-    model = GenerativeModel("gemini-1.5-flash-002")
+    model = GenerativeModel("gemini-2.5-pro")
     generate_content(
         model,
         [
-            Content(role="user", parts=[Part.from_text("Say this is a test")]),
+            Content(
+                role="user",
+                parts=[
+                    Part.from_text("Say this is a test"),
+                ],
+            ),
         ],
     )
 
     # Emits span
     spans = span_exporter.get_finished_spans()
     assert len(spans) == 1
-    assert spans[0].name == "chat gemini-1.5-flash-002"
+    assert spans[0].name == "chat gemini-2.5-pro"
     assert dict(spans[0].attributes) == {
         "gen_ai.operation.name": "chat",
-        "gen_ai.request.model": "gemini-1.5-flash-002",
+        "gen_ai.request.model": "gemini-2.5-pro",
         "gen_ai.response.finish_reasons": ("stop",),
-        "gen_ai.response.model": "gemini-1.5-flash-002",
+        "gen_ai.response.model": "gemini-2.5-pro",
         "gen_ai.system": "vertex_ai",
         "gen_ai.usage.input_tokens": 5,
-        "gen_ai.usage.output_tokens": 19,
+        "gen_ai.usage.output_tokens": 5,
         "server.address": "us-central1-aiplatform.googleapis.com",
         "server.port": 443,
     }
 
-    # Emits user and choice events
     logs = log_exporter.get_finished_logs()
+
+    # Emits user and choice events
     assert len(logs) == 2
     user_log, choice_log = [log_data.log_record for log_data in logs]
 
@@ -98,24 +89,20 @@ def test_generate_content(
         "finish_reason": "stop",
         "index": 0,
         "message": {
-            "content": [
-                {
-                    "text": "Okay, I understand.  I'm ready for your test.  Please proceed.\n"
-                }
-            ],
+            "content": [{"text": "This is a test."}],
             "role": "model",
         },
     }
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr()
 def test_generate_content_without_events(
     span_exporter: InMemorySpanExporter,
     log_exporter: InMemoryLogExporter,
-    generate_content: GenerateContentFixture,
+    generate_content: callable,
     instrument_no_content: VertexAIInstrumentor,
 ):
-    model = GenerativeModel("gemini-1.5-flash-002")
+    model = GenerativeModel("gemini-2.5-pro")
     generate_content(
         model,
         [
@@ -126,21 +113,21 @@ def test_generate_content_without_events(
     # Emits span
     spans = span_exporter.get_finished_spans()
     assert len(spans) == 1
-    assert spans[0].name == "chat gemini-1.5-flash-002"
+    assert spans[0].name == "chat gemini-2.5-pro"
     assert dict(spans[0].attributes) == {
         "gen_ai.operation.name": "chat",
-        "gen_ai.request.model": "gemini-1.5-flash-002",
+        "gen_ai.request.model": "gemini-2.5-pro",
         "gen_ai.response.finish_reasons": ("stop",),
-        "gen_ai.response.model": "gemini-1.5-flash-002",
+        "gen_ai.response.model": "gemini-2.5-pro",
         "gen_ai.system": "vertex_ai",
         "gen_ai.usage.input_tokens": 5,
-        "gen_ai.usage.output_tokens": 19,
+        "gen_ai.usage.output_tokens": 5,
         "server.address": "us-central1-aiplatform.googleapis.com",
         "server.port": 443,
     }
 
-    # Emits user and choice event without body.content
     logs = log_exporter.get_finished_logs()
+    # Emits user and choice event without body.content
     assert len(logs) == 2
     user_log, choice_log = [log_data.log_record for log_data in logs]
     assert user_log.attributes == {
@@ -160,10 +147,10 @@ def test_generate_content_without_events(
     }
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr()
 def test_generate_content_empty_model(
     span_exporter: InMemorySpanExporter,
-    generate_content: GenerateContentFixture,
+    generate_content: callable,
     instrument_with_content: VertexAIInstrumentor,
 ):
     model = GenerativeModel("")
@@ -193,10 +180,10 @@ def test_generate_content_empty_model(
     assert_span_error(spans[0])
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr()
 def test_generate_content_missing_model(
     span_exporter: InMemorySpanExporter,
-    generate_content: GenerateContentFixture,
+    generate_content: callable,
     instrument_with_content: VertexAIInstrumentor,
 ):
     model = GenerativeModel("gemini-does-not-exist")
@@ -226,13 +213,13 @@ def test_generate_content_missing_model(
     assert_span_error(spans[0])
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr()
 def test_generate_content_invalid_temperature(
     span_exporter: InMemorySpanExporter,
-    generate_content: GenerateContentFixture,
+    generate_content: callable,
     instrument_with_content: VertexAIInstrumentor,
 ):
-    model = GenerativeModel("gemini-1.5-flash-002")
+    model = GenerativeModel("gemini-2.5-pro")
     try:
         # Temperature out of range causes error
         generate_content(
@@ -249,10 +236,10 @@ def test_generate_content_invalid_temperature(
 
     spans = span_exporter.get_finished_spans()
     assert len(spans) == 1
-    assert spans[0].name == "chat gemini-1.5-flash-002"
+    assert spans[0].name == "chat gemini-2.5-pro"
     assert dict(spans[0].attributes) == {
         "gen_ai.operation.name": "chat",
-        "gen_ai.request.model": "gemini-1.5-flash-002",
+        "gen_ai.request.model": "gemini-2.5-pro",
         "gen_ai.request.temperature": 1000.0,
         "gen_ai.system": "vertex_ai",
         "server.address": "us-central1-aiplatform.googleapis.com",
@@ -261,13 +248,13 @@ def test_generate_content_invalid_temperature(
     assert_span_error(spans[0])
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr()
 def test_generate_content_invalid_role(
     log_exporter: InMemoryLogExporter,
-    generate_content: GenerateContentFixture,
+    generate_content: callable,
     instrument_with_content: VertexAIInstrumentor,
 ):
-    model = GenerativeModel("gemini-1.5-flash-002")
+    model = GenerativeModel("gemini-2.5-pro")
     try:
         # Fails because role must be "user" or "model"
         generate_content(
@@ -285,11 +272,12 @@ def test_generate_content_invalid_role(
     # Emits the faulty content which caused the request to fail
     logs = log_exporter.get_finished_logs()
     assert len(logs) == 1
-    assert logs[0].log_record.attributes == {
+    log = logs[0].log_record
+    assert log.attributes == {
         "gen_ai.system": "vertex_ai",
         "event.name": "gen_ai.user.message",
     }
-    assert logs[0].log_record.body == {
+    assert log.body == {
         "content": [{"text": "Say this is a test"}],
         "role": "invalid_role",
     }
@@ -299,7 +287,7 @@ def test_generate_content_invalid_role(
 def test_generate_content_extra_params(
     span_exporter,
     instrument_no_content,
-    generate_content: GenerateContentFixture,
+    generate_content: callable,
 ):
     generation_config = GenerationConfig(
         top_k=2,
@@ -311,7 +299,7 @@ def test_generate_content_extra_params(
         frequency_penalty=1.0,
         seed=12345,
     )
-    model = GenerativeModel("gemini-1.5-flash-002")
+    model = GenerativeModel("gemini-2.5-pro")
     generate_content(
         model,
         [
@@ -326,16 +314,16 @@ def test_generate_content_extra_params(
         "gen_ai.operation.name": "chat",
         "gen_ai.request.frequency_penalty": 1.0,
         "gen_ai.request.max_tokens": 5,
-        "gen_ai.request.model": "gemini-1.5-flash-002",
+        "gen_ai.request.model": "gemini-2.5-pro",
         "gen_ai.request.presence_penalty": -1.5,
         "gen_ai.request.stop_sequences": ("\n\n\n",),
         "gen_ai.request.temperature": 0.20000000298023224,
         "gen_ai.request.top_p": 0.949999988079071,
         "gen_ai.response.finish_reasons": ("length",),
-        "gen_ai.response.model": "gemini-1.5-flash-002",
+        "gen_ai.response.model": "gemini-2.5-pro",
         "gen_ai.system": "vertex_ai",
         "gen_ai.usage.input_tokens": 5,
-        "gen_ai.usage.output_tokens": 5,
+        "gen_ai.usage.output_tokens": 0,
         "server.address": "us-central1-aiplatform.googleapis.com",
         "server.port": 443,
     }
@@ -353,46 +341,46 @@ def assert_span_error(span: ReadableSpan) -> None:
     assert error_events != []
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr()
 def test_generate_content_all_events(
     log_exporter: InMemoryLogExporter,
-    generate_content: GenerateContentFixture,
+    generate_content: callable,
     instrument_with_content: VertexAIInstrumentor,
 ):
     generate_content_all_input_events(
         GenerativeModel(
-            "gemini-1.5-flash-002",
+            "gemini-2.5-pro",
             system_instruction=Part.from_text(
                 "You are a clever language model"
             ),
         ),
-        generate_content,
         log_exporter,
+        instrument_with_content,
     )
 
 
-@pytest.mark.vcr
+@pytest.mark.vcr()
 def test_preview_generate_content_all_input_events(
     log_exporter: InMemoryLogExporter,
-    generate_content: GenerateContentFixture,
+    generate_content: callable,
     instrument_with_content: VertexAIInstrumentor,
 ):
     generate_content_all_input_events(
         PreviewGenerativeModel(
-            "gemini-1.5-flash-002",
+            "gemini-2.5-pro",
             system_instruction=Part.from_text(
                 "You are a clever language model"
             ),
         ),
-        generate_content,
         log_exporter,
+        instrument_with_content,
     )
 
 
 def generate_content_all_input_events(
     model: GenerativeModel | PreviewGenerativeModel,
-    generate_content: GenerateContentFixture,
     log_exporter: InMemoryLogExporter,
+    instrument_with_content: VertexAIInstrumentor,
 ):
     model.generate_content(
         [
@@ -409,6 +397,9 @@ def generate_content_all_input_events(
                 ],
             ),
         ],
+        generation_config=GenerationConfig(
+            seed=12345, response_mime_type="text/plain"
+        ),
     )
 
     # Emits a system event, 2 users events, an assistant event, and the choice (response) event
@@ -463,57 +454,7 @@ def generate_content_all_input_events(
         "finish_reason": "stop",
         "index": 0,
         "message": {
-            "content": [{"text": "OpenTelemetry, this is a test.\n"}],
+            "content": [{"text": "OpenTelemetry, this is a test."}],
             "role": "model",
         },
     }
-
-
-_P = ParamSpec("_P")
-_R = TypeVar("_R")
-
-
-def _copy_signature(
-    func_type: Callable[_P, _R],
-) -> Callable[
-    [Callable[..., Any]], Callable[Concatenate[GenerativeModel, _P], _R]
-]:
-    return lambda func: func
-
-
-# Type annotation for fixture to make LSP work properly
-class GenerateContentFixture(Protocol):
-    @_copy_signature(GenerativeModel.generate_content)
-    def __call__(self): ...
-
-
-@pytest.fixture(
-    name="generate_content",
-    params=(
-        pytest.param(False, id="sync"),
-        pytest.param(True, id="async"),
-    ),
-)
-def fixture_generate_content(
-    request: pytest.FixtureRequest,
-    vcr: VCR,
-) -> Generator[GenerateContentFixture, None, None]:
-    """This fixture parameterizes tests that use it to test calling both
-    GenerativeModel.generate_content() and GenerativeModel.generate_content_async().
-    """
-    is_async: bool = request.param
-
-    if is_async:
-        # See
-        # https://github.com/googleapis/python-aiplatform/blob/cb0e5fedbf45cb0531c0b8611fb7fabdd1f57e56/google/cloud/aiplatform/initializer.py#L717-L729
-        _set_async_rest_credentials(credentials=AsyncAnonymousCredentials())
-
-    def wrapper(model: GenerativeModel, *args, **kwargs) -> None:
-        if is_async:
-            return asyncio.run(model.generate_content_async(*args, **kwargs))
-        return model.generate_content(*args, **kwargs)
-
-    with vcr.use_cassette(
-        request.node.originalname, allow_playback_repeats=True
-    ):
-        yield wrapper