+++ util/opentelemetry-util-genai/src/opentelemetry/util/genai/environment_variables.py
@@ -15,3 +15,31 @@
 OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT = (
     "OTEL_INSTRUMENTATION_GENAI_CAPTURE_MESSAGE_CONTENT"
 )
+
+OTEL_INSTRUMENTATION_GENAI_UPLOAD_HOOK = (
+    "OTEL_INSTRUMENTATION_GENAI_UPLOAD_HOOK"
+)
+"""
+.. envvar:: OTEL_INSTRUMENTATION_GENAI_UPLOAD_HOOK
+"""
+
+OTEL_INSTRUMENTATION_GENAI_UPLOAD_BASE_PATH = (
+    "OTEL_INSTRUMENTATION_GENAI_UPLOAD_BASE_PATH"
+)
+"""
+.. envvar:: OTEL_INSTRUMENTATION_GENAI_UPLOAD_BASE_PATH
+
+An :func:`fsspec.open` compatible URI/path for uploading prompts and responses. Can be a local
+path like ``/path/to/prompts`` or a cloud storage URI such as ``gs://my_bucket``. For more
+information, see
+
+* `Instantiate a file-system
+  <https://filesystem-spec.readthedocs.io/en/latest/usage.html#instantiate-a-file-system>`_ for supported values and how to
+  install support for additional backend implementations.
+* `Configuration
+  <https://filesystem-spec.readthedocs.io/en/latest/features.html#configuration>`_ for
+  configuring a backend with environment variables.
+*  `URL Chaining
+   <https://filesystem-spec.readthedocs.io/en/latest/features.html#url-chaining>`_ for advanced
+   use cases.
+"""